app/requirements.txt
Flask==3.0.2
Flask-JWT-Extended==4.6.0
Flask-SQLAlchemy==3.1.1
Flask-PyMongo==2.3.0
Flask-WTF==1.2.1
fonttools==4.47.0
gevent==24.2.1
gunicorn==21.2.0
ipython==8.20.0
Jinja2==3.1.2
nltk==3.8.1
openai==1.11.1
pandas==2.1.4
PyJWT==2.8.0
python-dotenv==1.0.1
PyPDF2==3.0.1
pymongo==4.6.1
mysql==0.0.3
mysqlclient==2.2.4
requests==2.31.0
scikit-learn==1.3.2
SQLAlchemy==2.0.27
summa==1.2.0

Werkzeug==3.0.1
WTForms==3.1.2
